# muenra-vouching/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Muenra Vouching API",
  "description": "\n**Muenra Vouching** — plataforma de vouching para auditoría financiera.\n\nFlujo: crear proyecto → importar Excel de referencia → cargar documentos →\nprocesar (XML / PDF / OCR / IA opcional) → motor de coincidencias → revisión humana → exportación.\n\nAutenticación: `POST /api/auth/login` devuelve un token Bearer (JWT).\n",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/projects/{project_id}/imports": {
   "post": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "Import Reference",
    "operationId": "import_reference_api_projects__project_id__imports_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Project Id"
      }
     },
     {
      "name": "dry_run",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Solo validar y devolver el informe de integridad",
       "default": false,
       "title": "Dry Run"
      },
      "description": "Solo validar y devolver el informe de integridad"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_reference_api_projects__project_id__imports_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "List Imports",
    "operationId": "list_imports_api_projects__project_id__imports_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}/references": {
   "get": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "List References",
    "operationId": "list_references_api_projects__project_id__references_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}/documents": {
   "post": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "Upload Documents",
    "operationId": "upload_documents_api_projects__project_id__documents_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Project Id"
      }
     },
     {
      "name": "process",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Encolar el procesamiento inmediatamente",
       "default": true,
       "title": "Process"
      },
      "description": "Encolar el procesamiento inmediatamente"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_documents_api_projects__project_id__documents_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "List Documents",
    "operationId": "list_documents_api_projects__project_id__documents_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Project Id"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}/process": {
   "post": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "Process Pending",
    "description": "Encola documentos pendientes (y opcionalmente con error) y una ejecución del motor de vouching.",
    "operationId": "process_pending_api_projects__project_id__process_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Project Id"
      }
     },
     {
      "name": "reprocess_errors",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Reprocess Errors"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}/match": {
   "post": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "Run Match",
    "operationId": "run_match_api_projects__project_id__match_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}/jobs": {
   "get": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "Jobs Status",
    "operationId": "jobs_status_api_projects__project_id__jobs_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/documents/{document_id}": {
   "get": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "Document Detail",
    "operationId": "document_detail_api_documents__document_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Document Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "Delete Document",
    "operationId": "delete_document_api_documents__document_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Document Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/documents/{document_id}/pages/{page}/image": {
   "get": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "Page Image",
    "operationId": "page_image_api_documents__document_id__pages__page__image_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Document Id"
      }
     },
     {
      "name": "page",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Page"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/documents/{document_id}/file": {
   "get": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "Download File",
    "operationId": "download_file_api_documents__document_id__file_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Document Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/documents/{document_id}/reprocess": {
   "post": {
    "tags": [
     "Documentos e importación"
    ],
    "summary": "Reprocess",
    "operationId": "reprocess_api_documents__document_id__reprocess_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Document Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Body_import_reference_api_projects__project_id__imports_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_import_reference_api_projects__project_id__imports_post"
   },
   "Body_upload_documents_api_projects__project_id__documents_post": {
    "properties": {
     "files": {
      "items": {
       "type": "string",
       "contentMediaType": "application/octet-stream"
      },
      "type": "array",
      "title": "Files"
     }
    },
    "type": "object",
    "required": [
     "files"
    ],
    "title": "Body_upload_documents_api_projects__project_id__documents_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}